# fix N+1 query in section auth