Further investigation in the student profile shows we were getting N+1 queries in the Section assignments authorization. This passes sections to the authorizer with the associated models needed to avoid this problem. Perf tests show significant improvement with this:

**authorized {Section.all}**
**Current**
Median: 24 seconds
P95: 35 seconds

**authorized {Section.all.includes(:course)}**
Median: 3 seconds
P95: 14 seconds

This reverts #2853 as well, since that improvement is no longer needed and is more complex that the original behavior.